Moving “tabs” module into “generic” sub-directory.
