+++ src/generic/tabs/Tabs.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { Dropdown } from '@edx/paragon';
 import { FormattedMessage } from '@edx/frontend-platform/i18n';
 import classNames from 'classnames';
+
 import useIndexOfLastVisibleChild from './useIndexOfLastVisibleChild';
 
 export default function Tabs({ children, className, ...attrs }) {
